Simple UI application in JavaScript.
Action items are usually documented in the meeting minutes and are recorded in the task list of the group. As people, complete action items, the items are documented as being completed and the item is removed from the list of outstanding action items.
This is what we tried to develop at the initial stage.